# docs/api-docs.yaml
openapi: 3.0.1
info:
  title: Bank
  description: Тестовое задание
  contact:
    name: Ветрова Анастасия
    email: Vetrova600@gmail.com
  version: v1
servers:
- url: https://localhost:8443
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/card-lock/request:
    post:
      tags:
      - Блокировка/разблокировка
      summary: Запрос на блокировку/разблокировку карты (USER)
      description: |
        Позволяет пользователю отправить запрос на смену статуса карты (например, заблокировать её).
        Передавайте `userId`, `cardId` и новый `cardStatus`(принимает значения "ACTIVE", "BLOCKED").
        Если карта уже находится в заданном статусе — возвращается ошибка.
      operationId: requestCardLock
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardLockOrderDto"
        required: true
      responses:
        "200":
          description: Запрос на изменение статуса карты успешно создан
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "400":
          description: Некорректные данные запроса или карта уже имеет указанный статус
          content:
            text/plain:
              schema:
                type: string
        "403":
          description: Недостаточно прав доступа
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Карта не найдена или не принадлежит пользователю
          content:
            text/plain:
              schema:
                type: string
...
  /api/v1/users:
    patch:
      tags:
      - Пользователи
      summary: Обновить данные пользователя (ADMIN)
      description: |
        Обновляет информацию о пользователе по переданному идентификатору.
        Только администратор имеет доступ к этому методу.
        Обновление происходит частично: поля, отсутствующие в запросе, остаются без изменений.
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "204":
          description: Данные пользователя успешно обновлены
        "400":
          description: Некорректные данные запроса
        "403":
          description: Недостаточно прав доступа
        "404":
          description: Пользователь не найден
  /api/v1/users/{userId}/role/{role}:
    delete:
      tags:
      - Пользователи
      summary: Удалить роль у пользователя (ADMIN)
      description: |
        Удаляет указанную роль у пользователя с заданным ID ("USER" или "ADMIN").
        Требуется наличие прав администратора.
        Если у пользователя нет такой роли — операция всё равно завершается успешно.
      operationId: removeRole
      parameters:
      - name: userId
        in: path
        description: ID пользователя
        required: true
        schema:
          type: string
          format: uuid
      - name: role
        in: path
        description: "Роль, которую нужно назначить"
        required: true
        schema:
          type: string
          enum:
          - USER
          - ADMIN
      responses:
        "204":
          description: Роль успешно удалена
        "403":
          description: Недостаточно прав доступа
        "404":
          description: Пользователь не найден
    patch:
      tags:
      - Пользователи
      summary: Назначить роль пользователю (ADMIN)
      description: |
        Назначает указанную роль пользователю по ID ("USER" или "ADMIN").
        Требуется наличие прав администратора.
        Повторное назначение существующей роли не влияет на результат.
      operationId: addRole
      parameters:
      - name: userId
        in: path
        description: ID пользователя
        required: true
        schema:
          type: string
          format: uuid
      - name: role
        in: path
        description: "Роль, которую нужно назначить"
        required: true
        schema:
          type: string
          enum:
          - USER
          - ADMIN
      responses:
        "204":
          description: Роль успешно назначена
        "400":
          description: Некорректные данные или роль уже назначена
        "403":
          description: Недостаточно прав доступа
        "404":
          description: Пользователь не найден
  /api/v1/users/first-admin:
    patch:
      tags:
      - Пользователи
      summary: Назначить первого администратора
      description: |
        Назначает текущему пользователю роль ADMIN,
        если в системе ещё нет пользователей с этой ролью.
        Повторное выполнение запрещено.
      operationId: createFirstAdmin
      responses:
        "204":
          description: Роль ADMIN успешно назначена
        "400":
          description: Роль ADMIN уже назначена другому пользователю
        "401":
          description: Пользователь не аутентифицирован
...
  /api/v1/card-lock/response:
    patch:
      tags:
      - Блокировка/разблокировка
      summary: Завершить заказ на блокировку карты (ADMIN)
      description: |
        Закрывает заказ на блокировку карты.
        Заказ должен находиться в статусе `IN_PROGRESS`, в противном случае будет выброшено исключение.
        Доступно только для пользователей с ролью ADMIN.
      operationId: closeOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseLockOrderDto"
        required: true
      responses:
        "204":
          description: Заказ успешно закрыт
        "400":
          description: Некорректный статус заказа или невалидные данные
        "403":
          description: Недостаточно прав доступа
        "404":
          description: Заказ не найден
  /api/v1/card-lock/order:
    patch:
      tags:
      - Блокировка/разблокировка
      summary: Получить задание на обработку блокировки карты (ADMIN)
      description: |
        Возвращает заказ на блокировку карты для администратора.
        Если `orderId` не передан, выбирается первый доступный заказ со статусом `PENDING`.
        Заказ переводится в статус `IN_PROGRESS` и привязывается к текущему администратору.
        Доступно только для пользователей с ролью ADMIN.
      operationId: getOrder
      parameters:
      - name: orderId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Заказ на блокировку карты успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardLockProcessingResponse"
        "403":
          description: Недостаточно прав доступа
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Заказ не найден
          content:
            text/plain:
              schema:
                type: string
...
components:
  schemas:
...
    CardLockOrderDto:
      required:
      - cardId
      - reason
      - requestedStatus
      type: object
      properties:
        cardId:
          type: string
          description: "ID карты, для которой меняется статус"
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        requestedStatus:
          type: string
          description: Статус банковской карты
          example: BLOCKED
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        reason:
          maxLength: 255
          minLength: 0
          type: string
          description: Причина блокировки/разблокировки
          example: Пользователь сообщил о краже карты
      description: Запрос на изменение статуса карты (блокировка/разблокировка)
...
    UpdateUserRequest:
      required:
      - id
      type: object
      properties:
        id:
          type: string
          description: ID пользователя
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Имя пользователя
          example: Иван
        surname:
          type: string
          description: Фамилия пользователя
          example: Иванов
        phone:
          pattern: "\\+7\\d{10}"
          type: string
          description: Телефон пользователя
          example: "+79001234567"
      description: |-
        **Передавайте только ID и поля, которые хотите изменить.**

        Запрос на обновление данных пользователя.
...
    CloseLockOrderDto:
      required:
      - id
      - status
      type: object
      properties:
        id:
          type: integer
          description: ID заявки на блокировку
          format: int64
          example: 1
        status:
          type: string
          description: Статус заказа
          example: APPROVED
          enum:
          - PENDING
          - IN_PROGRESS
          - APPROVED
          - REJECTED
        adminComment:
          maxLength: 255
          minLength: 0
          type: string
          description: Комментарий администратора
          example: "Проверка завершена, блокировка одобрена"
      description: DTO для закрытия заявки на блокировку карты
    CardLockProcessingResponse:
      type: object
      properties:
        id:
          type: integer
          description: ID заявки
          format: int64
          example: 1
        userId:
          type: string
          description: "ID пользователя, подавшего заявку"
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        cardId:
          type: string
          description: "ID карты, которой нужно поменять статус"
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        adminId:
          type: string
          description: "ID администратора, обработавшего заявку"
          format: uuid
          nullable: true
          example: 123e4567-e89b-12d3-a456-426614174000
        requestedStatus:
          type: string
          description: Статус банковской карты
          example: BLOCKED
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        reason:
          type: string
          description: Причина изменения статуса
          example: Подозрение на мошенничество
        createdAt:
          type: string
          description: Дата создания заявки
          format: date
          example: 2025-06-27
        adminComment:
          type: string
          description: Комментарий администратора
          nullable: true
          example: Подтверждено
        status:
          type: string
          description: Статус заказа
          example: IN_PROGRESS
          enum:
          - PENDING
          - IN_PROGRESS
          - APPROVED
          - REJECTED
      description: Информация о заявке на блокировку/разблокировку карты
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT